requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.30.0
anyio==4.4.0
arxiv==2.1.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.6
datasets==2.20.0
decorator==5.1.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
evaluate==0.4.2
exceptiongroup==1.2.1
executing==2.0.1
feedparser==6.0.11
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface==0.0.1
huggingface-hub==0.23.4
idna==3.7
ipython==8.12.3
jedi==0.19.1
jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
lightning-utilities==0.11.3.post0
lit==18.1.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.1
numpy==1.24.4
openai==1.35.6
packaging==24.1
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.2
prompt-toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.4
pydantic-core==2.18.4
pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.3.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.0.1
sentry-sdk==2.13.0
setproctitle==1.3.3
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
stack-data==0.6.3
sympy==1.12.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch
torchaudio
torchmetrics
torchvision
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.0
triton==3.0.0
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.7
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
accelerate
